requirements.txt
Django==5.2.1
asgiref==3.8.1
sqlparse==0.5.3

# Production Database
psycopg2-binary==2.9.7

# AWS Deployment
gunicorn==21.2.0
boto3==1.34.34
django-storages==1.14.2

# Analytics and Visualization
django-extensions==3.2.3
plotly==5.17.0
pandas==2.1.4

# Environment Configuration
python-decouple==3.8

# Logging and Monitoring
django-debug-toolbar==4.2.0

# Security
whitenoise==6.6.0
